# src/web/openapi/running-performance.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RunningPerformance.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/activities": {
      "get": {
        "tags": [
          "Activities"
        ],
        "operationId": "GetActivities",
        "parameters": [
          {
            "name": "activityType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "modality",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityPageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/activities/{id}": {
      "get": {
        "tags": [
          "Activities"
        ],
        "operationId": "GetActivity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivityDetailResponse": {
        "required": [
          "activity",
          "startedAtUtc",
          "timezoneName",
          "utcOffsetMinutes",
          "movingSeconds",
          "elapsedSeconds",
          "averageSpeedMps",
          "calories",
          "averageCadenceSpm",
          "averagePowerW",
          "elevationGainM",
          "lapCount",
          "sources"
        ],
        "type": "object",
        "properties": {
          "activity": {
            "$ref": "#/components/schemas/ActivitySummaryResponse"
          },
          "startedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "timezoneName": {
            "type": [
              "null",
              "string"
            ]
          },
          "utcOffsetMinutes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int16"
          },
          "movingSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "elapsedSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "averageSpeedMps": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "calories": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "averageCadenceSpm": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "averagePowerW": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "elevationGainM": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "lapCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivitySourceResponse"
            }
          }
        }
      },
      "ActivityPageResponse": {
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivitySummaryResponse"
            }
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "ActivitySourceResponse": {
        "required": [
          "id",
          "sourceClass",
          "sourceRowNumber",
          "linkingResult",
          "observedAt",
          "summary",
          "sourceFileId",
          "originalName",
          "sha256",
          "ingestionRunId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sourceClass": {
            "type": "string"
          },
          "sourceRowNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "linkingResult": {
            "type": [
              "null",
              "string"
            ]
          },
          "observedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "summary": {
            "$ref": "#/components/schemas/JsonElement"
          },
          "sourceFileId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "originalName": {
            "type": [
              "null",
              "string"
            ]
          },
          "sha256": {
            "type": [
              "null",
              "string"
            ]
          },
          "ingestionRunId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
      "ActivitySummaryResponse": {
        "required": [
          "id",
          "provisionalActivityKey",
          "garminActivityId",
          "activityType",
          "activityCategory",
          "modality",
          "startedAtLocal",
          "title",
          "distanceM",
          "durationSeconds",
          "averagePaceSecondsPerKm",
          "averageHeartRateBpm",
          "maxHeartRateBpm",
          "validationStatus"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "provisionalActivityKey": {
            "type": [
              "null",
              "string"
            ]
          },
          "garminActivityId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "activityType": {
            "type": "string"
          },
          "activityCategory": {
            "type": [
              "null",
              "string"
            ]
          },
          "modality": {
            "type": [
              "null",
              "string"
            ]
          },
          "startedAtLocal": {
            "type": "string",
            "format": "date-time"
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "distanceM": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "durationSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "averagePaceSecondsPerKm": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "averageHeartRateBpm": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "maxHeartRateBpm": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "validationStatus": {
            "type": "string"
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "JsonElement": { },
...
    }
  },
...
}